Starts sql type consolidation
