+++ opsml/app/routes/files.py
@@ -26,7 +26,7 @@ from opsml.app.routes.utils import (
     MaxBodySizeValidator,
 )
 from opsml.helpers.logging import ArtifactLogger
-from opsml.registry.sql.table_names import RegistryTableNames
+from opsml.registry import RegistryTableNames
 
 logger = ArtifactLogger.get_logger()
 CHUNK_SIZE = 31457280
