+++ opsml/registry/__init__.py
@@ -13,6 +13,7 @@ from opsml.registry.cards.types import (
 from opsml.registry.data.splitter import DataSplit
 from opsml.registry.sql.registry import CardRegistries, CardRegistry
 from opsml.registry.sql.semver import VersionType
+from opsml.registry.sql.types import RegistryTableNames
 
 __all__ = [
     "AuditCard",
@@ -29,4 +30,5 @@ __all__ = [
     "CardRegistries",
     "CardRegistry",
     "VersionType",
+    "RegistryTableNames",
 ]
