+++ opsml/registry/cards/base.py
@@ -11,7 +11,7 @@ from opsml.helpers.logging import ArtifactLogger
 from opsml.helpers.utils import clean_string, validate_name_team_pattern
 from opsml.registry.cards.types import CardInfo
 from opsml.registry.sql.records import RegistryRecord
-from opsml.registry.sql.table_names import RegistryTableNames
+from opsml.registry.sql.types import RegistryTableNames
 
 logger = ArtifactLogger.get_logger()
