+++ opsml/registry/sql/base/registry_base.py
@@ -15,7 +15,7 @@ from opsml.registry.cards.card_saver import save_card_artifacts
 from opsml.registry.cards.types import RegistryType
 from opsml.registry.sql.records import LoadedRecordType, load_record
 from opsml.registry.sql.semver import CardVersion, SemVerUtils, VersionType
-from opsml.registry.sql.table_names import RegistryTableNames
+from opsml.registry.sql.types import RegistryTableNames
 from opsml.registry.storage.settings import DefaultSettings
 
 logger = ArtifactLogger.get_logger()
